[BUGFIX] Fix validators messages translation fatal error in external unit tests
The function `translateErrorMessage()` of an Extbase validator would cause a
fatal error because of its usage of `LocalizationUtility` (which tries to access
TYPO3 database).

The solution was to intercept the validator creation in the mocked object
manager, and return a mocked validator instead.

It means that when an error is returned during the creation of a
configuration object, the message will not be translated, but look like this:
`LLL:$extension:$key` where `$extension` is the name of the extension
containing the translation, and `$key` the key of the translation reference.

The test `ValidatorResolverTest` had to be refactored a bit to be compatible
with the new way of working with validators during unit testing. It has no
incidence on the logic of the test.