[BUGFIX] Fix incorrect cache handling
This commit refactors a major part of how "dynamic" caches were handled
by Configuration Object API. In the provided cache service that can
be attached to a configuration object, it is possible to declare the
options for this cache; it means the cache is registered long after
TYPO3 initialization, resulting in issues like caches entries not being
deleted on caches flush.

A new internal cache service has been introduced, which will handle
these "dynamic" caches: when one of these caches is used, its
configuration is saved in the internal cache: in further request, the
cache will be properly registered during TYPO3 initialization.