[FEATURE] Introduce support for common array object annotation (#17)
Previously, when a property would be filled by an array of object, only two
annotations were supported:

`\ArrayObject<\Some\Class>` and `array<\Some\Class>`

They are not user-friendly, and most of the time the array is a basic array (not
an `ObjectStorage` for instance). The common annotation for this would be:

`\Some\Class[]` - which provides better IDE auto-completion, and has better
readability.

This commit introduces support for this annotation, enjoy!