[TASK] Return mapping result if it contains errors
When the mapping result given as constructor argument of a configuration object
instance contains errors, it will always be returned when the validation result
is accessed (function `getValidationResult()`).

This is done because if the mapper did not succeed to build a proper object and
its sub-objects, then it makes no sense in many cases to launch a whole
validation process on an unfinished object.

This commit leads to more consistency in an object validation result messages.