[TASK] Return error when required constructor argument is missing
When an object is built with the object converter, if a required argument for
the constructor is missing, it will now return an error instead of throwing an
exception.

This way, it will be added to the mapping result, and will let the user have
more information about why the object is not valid.