[TASK] Filter parents before they are attached to an object
For a given object using the `ParentsTrait`, its parent objects are now filtered
by removing unnecessary parents that use `ParentsTrait` and are followed by
another parent that does the same: there is no need to have the full list of
chained parents.

This does not change the implementation: this will just lighten the array used
to store the parents.