[FEATURE] Introduce `ParentsTrait::alongParents()` function
This function allows to call a callback function for each parent of an object
that uses `ParentsTrait`.

The callback function has a single parameter which is the current parent object.
If `false` is returned by the callback, the loop on the parents stops.

This new function is now used by the following functions of `ParentsTrait`:
`hasParent`, `withFirstParent` and `getFirstParent`.