[FEATURE] Introduce configuration object factory `isRunning` method
A new method has been added: `ConfigurationObjectFactory::isRunning()`.

With this method, you can check at any moment if the configuration object
factory is currently processing (an object is being created). This can be useful
for instance if you want to allow magic methods for an object only when it is
being converted.