[BUGFIX] Handle mixed types for properties with single object
Object properties that used the "mixed types" feature were not checked
if the property was filled with a single instance. Only composite types
(means the properties contains several object instances) were actually
checked.

This commit patches the validator resolver (in a 💩 way) to allow it to
handle single instances. A new internal validator has been introduced:
it will wait for the actual object instance to be validated, in order to
create an up to date validator conjunction.